Revert "Merge remote-tracking branch 'upstream/klone-MK3' into clone-MK25S"
This reverts commit 0da5d1b0f739b32240f227a0a7ef66fa6fe4d185, reversing
changes made to 47b2e690df72015c24d05a8a9bd2430dab5334b5.

+++ Firmware/conv2str.cpp
@@ -104,7 +104,7 @@ char *ftostr43(const float &x, uint8_t offset)
 //Float to string with 1.23 format
 char *ftostr12ns(const float &x)
 {
-  int xx = x * 100;
+  long xx = x * 100;
 
   xx = abs(xx);
   conv[0] = (xx / 100) % 10 + '0';
